backend/requirements.txt
alembic==1.11.1
blinker==1.6.2
cffi==1.15.1
click==8.1.3
cryptography==40.0.2
Flask==2.3.2
Flask-Cors==3.0.10
greenlet==2.0.2
gunicorn==20.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.2
marshmallow==3.19.0
packaging==23.1
passlib==1.7.4
psycopg2-binary==2.9.6
pycparser==2.21
PyJWT==2.7.0
python-dotenv==1.0.0
six==1.16.0
SQLAlchemy==2.0.15
typing_extensions==4.6.1
Werkzeug==2.3.4
